pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "keys-r-us"
description = ''
readme = "README.md"
requires-python = ">=3.7"
license = "MIT"
keywords = []
authors = [
    { name = "Evan Aranda", email = "evanaranda35@gmail.com" },
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
    "mnemonic",
    "shamir-mnemonic",
    "base58",
    "click",
    "qrcode",
    "pyzbar",
    "pillow"
]
dynamic = ["version"]

[project.urls]
Documentation = "https://github.com/unknown/keys-r-us#readme"
Issues = "https://github.com/unknown/keys-r-us/issues"
Source = "https://github.com/unknown/keys-r-us"

[project.scripts]
keys-r-us = "keys_r_us.cli:entry"

[tool.hatch.version]
path = "keys_r_us/__about__.py"

[tool.hatch.envs.default]
path = ".hatch"
python = "python3.11"
dependencies = [
    "pytest",
    "pytest-cov",
]
[tool.hatch.envs.default.scripts]
cov = "pytest --cov-report=term-missing --cov-config=pyproject.toml --cov=keys_r_us --cov=tests {args}"
no-cov = "cov --no-cov {args}"

[[tool.hatch.envs.test.matrix]]
python = ["37", "38", "39", "310", "311"]

[tool.coverage.run]
branch = true
parallel = true
omit = [
    "keys_r_us/__about__.py",
]

[tool.coverage.report]
exclude_lines = [
    "no cov",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
